Unbuffer multiraft's input channels.
Buffering at this level makes it difficult to guarantee sequencing of
different operations from the outside.

This also fixes a potential bug in re-proposals: if the pending queue
ever exceeded the size of proposalChan's buffer, we would deadlock
trying to write to proposalChan. We now avoid this problem by calling
state.propose directly instead of sending everything back through the
channel.