merge clog and fix flag handling
decided to merge clog into util/log since the flag handling was ridiculous
and we were only using clog as an engine for efficient writing of log output
anyways. Unexported a bunch of things and added in code that allows the use
of log flags for the tests. This meant adding a few lines of boilerplate to
each package, but it doesn't seem too bad. Also fixed up the clog code for
`make check`.

Also changed the default of --logtostderr to true. If one wants to log to
files only, --log-dir=/some/dir and --logtostderr=false (and presumably
--stderrthreshold=<some_number> should also be specified).