Merge allocator and store finder; remove store finder.
Implemented two methods on allocator: Allocate() and Rebalance().
Rebalance() is what we'll use most of the time and always chooses
a random store which will aid in rebalancing. Allocate() chooses
from all stores with a bias towards ones with more capacity. This
will be used when recovering from dead stores/nodes.