Added an error margin for range count based rebalance decisions.
Changed RebalanceTarget() to use AllocateTarget() more exactly. That is,
we now choose three random targets from all available stores which match
the attributes, and find the best result which matches rebalancing criteria.

The difference is that now we will be less likely to find a target
successfully in the event that there are enough potential targets which
aren't rebalanceable. However, this is actually OK, because the ratio
of balancers to balancees for the particular class of store evens the
odds so we expect balancings to continue at the correct pace without
overcrowding any particular target or targets.