Convert Admin UI to MithrilJS
This checkin converts the existing administrative UI pages to use the Mithril JS
framework instead of Angular. Mithril is a compact and relatively young
framework for single-page javascript apps.

We chose Mithril for a variety of reasons:
+ Mithril was chosen over Angular primarily for its rendering speed, facilitated
by the use of a Virtual DOM (similar to the React framework). It is also a
considerably simple framework.
+ Mithril was chosen over React because it is immediately compatible with
Typescript, and is a much smaller package to embed.

The RestExplorer UI, though not often used, has been converted as part of this
change; this was mostly an effort to understand how to best organize Mithril
components in Typescript.