Fix handling of "infinite" clock offsets.
Infinite clock offsets are generated when a heartbeat times (6s) out or
takes too long to finished (greater than 1.5s). Previously, such offsets
were treated as valid values as part of the cluster offset interval
calculation, but doing so meant that a sufficient number of slow nodes
in a cluster (or a sufficiently slow network) would guarantee an
unhealthy clock offset interval to be calculated.

Now we never add "infinite" offsets to the clock offset monitor map. The
monitor is more pessimistic and will only suicide if the set of
good (non-infinite) clock offsets it has computes a cluster offset
interval that is larger than the max offset interval. If a node cannot
talk to any other nodes then its cluster offset interval will be
considered healthy.