Changed logging to be human readable.
Added color highlighting if the term supports it for warning=red,
error=bold&red.

We now build up a dictionary for output to a binary log (TODO).
I removed ClusterID from the log Field enum. It's completely superfluous.

Also changed the Infoc, Warningc, Errorc & Fatalc to take actual
format string and parameters. The other mechanism was just too clumsy
to be used easily. As part of the machine-readable dict, we output
the original format string and each of the parameters to it, along
with their types. I feel like this should allow pretty flexible
programmatic manipulation of logs.

Presumably the machine readable logs will be output as JSON and compressed.