Fix binding of the certs directory when using boot2docker.
Boot2docker's handling of binding local directories into the container
is very confusing. If the directory being bound has a parent directory
that exists in the boot2docker VM then that directory is bound into the
container. In particular, that means that binds of /tmp and /var will be
problematic.