Speedup cluster creation by using pre-generated certs.
This reduces the total run time of the acceptance tests from 42sec to
27s on my machine.

Added TestCertGen which generates new certs and verifies the nodes can
talk to each other.

Rename the "-n" flag to "-num" to avoid a conflict with the "-n" flag in
the go tool.

Regenerated the test_certs to contain alternate names for
"roach{0-99}.local".