New 'IsRangeOp' flag on methods and better enforcement of keys
For ranged ops, the end key must be present and must be different
from the start key. For non-ranged ops, the end key must not be
specified.

Created a new InternalResolveIntentRange op to keep it separate from
the non-ranged version. Having them as the same op was getting
confusing and seemed to be inviting problems.