Add Time Series Rate Aggregator
Introduces the concept of an Aggregator function to time series queries. Time
series now have a choice of two aggregation functions, Average and Rate.

Average captures the previous behavior, where queries return the average value
of each sample period. Rate returns the rate of change for the series at each
sample period; the rate denominator is the length of the sample period
(currently 10 seconds for all metrics).

More work needs to be done to expose the sample period length in queries,
especially in the case of Rate; this should be considered a partially complete
work, which is desired to support upcoming product demos.