Refactor Graph Component
Refactor to add "QueryManager" component, which can helpfully persist data from
a query for use by multiple synchronized components.

Also included is an example of how a top level component (graph demo page) can
manage the refreshing of a QueryManager while passing it to lower-level
components for rendering.

Finally, fixes a bug in the Go code where the NodeID was not being set for the
status recorder, preventing the Node-specific stats from being recorded.