avoid redundant descriptor cache expiration
especially during a range split, multiple clients may receive retriable
errors indicating that their commands were sent to the wrong range.
the reaction to this is to evict the cached range descriptor, which
prompts an expensive query on the next retry.
however, it makes no sense to evict a range descriptor which is not
equal to the one taken out of the cache earlier. hence in this change
we optionally allow passing in the seen range descriptor when evicting,
which the cache uses to pass on the eviction if it believes that that
seen descriptor has been evicted previously anyways.