work around raft message reordering (#1176)
in #1176 we discovered that (on local networks) Raft messages may be
received in the wrong order. The likely culprit at the time of writing
is the receiver side, which spawns a goroutine for each incoming request,
allowing reordering (we have checked that client.Go is called in the right
order on the sending side, and that serializes outgoing messages).

The reordering doesn't cause inconsistencies but slows down the cluster.
This commit adds a temporary workaround until we fully address the issue.