Refactor Node/Store Admin Pages Pt. I
Initial set of refactors for node/store status pages in the admin UI:
+ Remove unused "historical status" functionality.
+ Commonizes most of the data manipulation functions used by Store and Node
models
+ Moves Store and Node models to a shared Status file

There is still refactoring to be done after this change; for example, there is
still view code in the model.  However, this change is already becoming
considerably complex.

This commit also fixes a bug in the Makefile where the Typescript was never
rebuilt.