Fix a race involving snapshots and the right-hand side of a split.
After a split, the newly-created range has multiple members (same as the
parent range) and can therefore complete elections and send raft
traffic, but the log index is reset. Since the range is considered
initialized, we need to reset the appliedIndex to raftInitialLogIndex
instead of 0 (which signifies an uninitialized range). If a range became
leader with an appliedIndex of 0, it could panic while trying to
generate a snapshot to send to a lagging follower.

Commit dd45ddda fixed a related issue involving the left-hand side of a
split.

Fixes #1191.
Fixes #1242.