more context.Context on receiver path
Node -> LocalSender -> Store -> Range [ -> Raft -> ] -> Range

is the chain along which the context.Context is now passed,
which gives us pretty good coverage of a client request once
it has been received at a Node (which is the opposite end
from the DistSender).
The context is not passed into Raft, but the replica which
proposed the command keeps track of it and passes the context
on to the application of the command; the other nodes will
run it with their default range-level context.

Replaced all log calls in store and range with their contextualized
counterparts.