Publish Replication Status to Store Feed
This commit is intended to bring the time series data maintained by the
server/status module up to parity with the data in the store status scanner.
This is in preparation for removing the store status scanner; we are going to
instead have the server/status module publish per-store and per-node status
summaries.

In order to accomplish this, we need the server/status module to be aware of
certain statistics that cannot currently be computed from incremental events
(availability/replication stats for ranges, OS storage capacity). In order to
accomplish this, we now periodically compute each store's replication and
capacity status, and publish that information as an event to the store's feed.

This allows the server/status module to pull replication and capacity
information from the feed. The replication information is now being stored as
time series data.

Note that this commit does NOT remove the store status scanner; the store
scanner is still responsible for generating status summaries. The scanner will
be dismantled in a future commit.