Fix TestServerNodeEventFeed on Multi-CPU
This test was encountering a race condition with an asynchronous background
task, which affected its expected results on Multi-CPU test runs.

Specifically, this test monitors the node event feed, recording all RPCs called
on the node. However, in addition to the RPCs called by the test, there are also
RPCs being called by the Node's "StoresScanner", which were showing up in the
event feed. The exact ordering was unusually stable in single CPU tests, but the
race is readily exposed in a multi-CPU test run.

This *temporary* fix relaxes the final assertion of the test; the test still
requires that it sees the events it generates directly, but now it will not fail
if there are additional events that it did not expect.

A better, permanent fix is suggested in the TODO.