Implement NodeStatusRecorder.GetStatusSummaries()
This method implements the summary generation that was previously implemented in
`store.updateStoreStatus()` and `node.startStoresScanner()`; however, the new
implementation uses the information collected by NodeStatusRecorder.

The method is intended to be called periodically by higher-level component,
which will be responsible for writing the values to the database. There is
currently nothing consuming them as of this commit.

The old scanner-based summary system is still intact as of this commit; however,
we have now created a viable replacement and the old system can be dismantled in
a subsequent change.