Remove `wait` parameter
The only case that exercised `wait=false` was a call to
`resolveWriteIntentError`, which turns out to have been incorrect; it
would return an error, but that error would be summarily ignored.

This commit removes the `wait` parameter, defaulting the behaviour of
affected methods to the `wait=true` case, with the exception of
`resolveWriteIntentError` which has been changed to do all of its
already best-effort work asynchronously in all cases.