Take longer naps in TestTxn*
Many of these tests cause some transactions to be repeatedly pushed.
Normally, those transactions would retry at the sender indefinitely,
but this test has places to be, so it limits the number of attempts
to make sure things finish in a timely manner.

Unfortunately this approach is inherently racy as it expects
transactions that are able to make progress in the context of their
history to not exhaust the retries allotted to them. If these
transactions retry faster than their pushers are able to commit, the
test will spuriously fail.

This commit increases backoff duration for these retries to increase
the probability that progress-capable transactins do not eagerly
exhaust their allotted retries.