Write Summaries from server/status package
This commit switches status summaries to be written from the server/status
package. As part of this, the old system of scanner-based stat collection has
been dismantled.

The most complicated part of this change is the update to server-level status
tests to accomodate the new system.

+ `TestNodeStatus` has been renamed to `TestStatusSummaries`
+ A test specific to StoreStatuses in "client_status_test" has been removed and
folded into `TestStatusSummaries`. This is possible because Node and Store
statuses are now written from the same place at the same time.
+ `CanaryEvent` has been added to the list of node status events, which can be
used to guarantee that the feed pipeline has been flushed by downstream
consumers. This is used to synchronize writes inside of `TestStatusSummaries`,
as opposed to the old system of waiting for scan completions.