Fix error handling for unknown methods in rpc codec.
Previously, an unknown method would leave the connection in an
inconsistent state so the next request could not be parsed
correctly (the connection would probably error out but it would be
possible to parse the next frame as the wrong type of protobuf).

Added some tests for new rpc server.