fix #1684
resolving intents is not necessary for a transaction to complete.
however, open intents which remain unresolved may prevent tasks
from completing if they are on meta records which preclude an
outstanding request from reaching the correct range, retrying
indefinitely.
This change refactors TxnCoordSender such that a new transaction
cannot be started when the system is draining, and moves the intent
cleanup into the teardown of the heartbeat goroutine, which is now
a task.

fixes #1684 since the issue there was precisely that: a split at
["db1", "db2"] was started, and the second split would never route
to the correct range unless the intents of the first were resolved
previously (which would sometimes not happen since the stopper had
already stopped accepting "new work").