Simplify Insert.Columns.
Previously Insert.Columns was a list of SelectExprs, but the parser only
ever generated QualifiedNames.

Took the opportunity to do various simplifications to the grammar
replacing a number of identical rules with a single rule (e.g. iconst ->
ICONST, sconst -> SCONST). This reduces the size of the generated code
slightly with no functional change.