Add basic WHERE clause support.
Basic in the sense that we simply evaluate the WHERE clause expression
for each row before the row is output. No analysis of the WHERE clause
is performed to determine if an index could be used or if additional
columns are necessary beyond those present in the select expressions.

Fixes #1624.