UI: Add Property and Computed Property to Utils
Add support for lazily computed properties to our library.

Computed properties depend on "explicit" properties, which have a similar
signature to mithril properties (a simple getter/setter function). However,
these properties also have an "Epoch()" value, which is a monotonically
increasing counter incremented whenever the property is set.

Computed properties are defined by a set of 'parent' properties and a function.
The function uses values of the parent properties to create a computed value.

The computed value is memoized whenever its value is accessed; however, if any
of the parent values have changed since the memoization (detected via the
parent's Epoch() values), the value is recomputed.

Computed properties also have an Epoch(), which is equal to the sum of its
parent Epochs. Computed properties can thus be chained together.

This commit also converts the existing "QueryCache" class to use Properties.