Further tweaks to cockroachdb/cockroach image building.
No longer build the tests and run them inside the deploy container. This
work was duplicative of the normal testing and served little
purpose. Removed the associated "testbuild" target from the Makefile.

Switched to using "busybox" as the base image. This reduces the
cockroachdb/cockroach image size from 210MB to 88MB.