fix a bug when splitting at a meta2 entry
fixes a bug which caused a panic when clearing overlapping descriptors for a
range [KeyMin, Meta2Key). The issue was that when attempting to clear out
descriptors which were sub- sumed by the above range, an iteration over the
range given by its meta keys was performed, with the first endpoint excluded.
This exclusion was performed incorrectly by first incrementing the start key
(KeyMin) and then forming the meta key; for KeyMin this leads to
Meta2Prefix\x00. For the above EndKey, the meta key is a Meta1key which sorts
before Meta2Prefix\x00, causing a panic. The fix was simply to increment the
meta key for StartKey, not StartKey itself.