Install lodash library
Lodash is an functional programming utility library, in the style of
underscore.js. It seems to be widely preferred over underscore at this point,
offering better performance and a larger (superset) API.

Another alternative considered was lazy.js, but it is apparently not under
active development anymore. Additionally, a great deal of lazy evaluation was
recently added to lodash, resulting in some of the same significant performance
increases that would have been available in lazy.js.

Includes typescript typings from DefinitelyTyped.