Typography structure in Stylus
Commit adds some support for better typography in our stylus/css:

+ Font sizes are now generated using Modular Scale
+ Root font size (rem unit) is configured so that we can easily generate a
vertical rhythm for the site. This based on the same principle as the "Sassline"
baseline grid system for SASS, although our usage is dramatically simpler.

tldr; All heights should be expressed in multiples of 2rem.  1 rem multiple can
be used for some precision placements if necessary.  By doing this, we establish
a consistent vertical rhythm for the page.