pin txn writes to a single coordinator
a subgoal of #1873 is making sure that a transaction can only write
on one coordinator (actually, the RFC proposes limiting reads to a
single coordinator as well, but this has proven unnecessary).

this change is carried out here. as a slight change from the RFC,
we do add an additional flag to the proto.Transaction struct, which
contains the information of whether previous requests in the Txn have
written data (i.e. laid down intents).
not doing that would have mandated giving up on the optimizations for
read-only transactions since we would have had to rely on special-
casing the first transactional request.