range-local intent resolution
this passes all tests except for one (which I've disabled; maybe
@bdarnell has an idea?). Basically carried out a lot of the work announced in
RFC #1873:

- coordinator sends intents along with EndTransaction
- range resolves all it possibly can in the EndTransaction batch, and hands
  everything else off to an async resolver.
- "skipped intents" passed up by the range_command functions are now executed
  even if the command returns an error. This is required for EndTransaction
  which finds out that it's been aborted.

There are a lot of TODOs documented in the code, and I intend to address all
of those that are relevant for correctness. Some parts of the RFC are also
not yet implemented (gc'ing Txn entries), but the road should be very clear
and I'll add those before merging.

I did want to get this out early because it's a fairly complex refactoring
and will only get bigger from here.