Adjustments to the SQL logic test framework.
Added support for a "colnames" query option which controls whether the
column names are printed at the start of query results. Defaults to
false since this is an extension of the sqllogictest format. Adjusted
the existing tests to use this where needed.

Created a default "test" database and set the session database to "test"
as part of the framework. The sqllogictest data files expect an
environment in which tables can be created without qualification.

Added a "-d" flag that allows changing which test data files are used.