Admin UI: Add Page Tab Navigation Bar
In the latest design wireframes, certain UI pages are composed of multiple
"Tabs" - for example, the page for a single Node will have an "Overview" tab and
a "Graphs" tab.

Navigation between these tabs is accomplished via a Navigation Bar component,
which is the same component used for the main navigation bar.

Finally, the content on the various page tabs has been restructured to more
closely resemble the current wireframes.