Remove parser.Env for expression evaluation.
Columns (qualified names) in expressions must now be replaced before
expression evaluation. This is accomplished via a WalkExpr visitor. This
allows for more sophisticated determination of the column that matches a
qualified name as well more efficient evaluation of the
expression. *QualifiedName nodes are replaced with *ParenExpr nodes, the
contents of which are filled in before the expression is evaluated.

Added support for expanding StarIndirection in QualifiedNames into a
list of columns when used in a select expression. For example:

  SELECT foo.* FROM foo

Added support for table aliases. For example:

  SELECT foo.* FROM bar AS foo

See #1986