Add support for "EXPLAIN (DEBUG)".
Added parser.Explain statement and support for generating this syntax
tree element to the grammar.

Added support for explain-debug to scanNode. This is currently the only
node which supports it. Lots of refactoring of scanNode occurred,
including breaking apart the unwieldy scanNode.Next().

Reworked how table aliasing is performed so that the original table name
is still available. Reworked how column visibility is limited during
scanning from an index in order to allow the full set of columns to be
available in explain-debug mode.

Reworked decodeIndexKey to avoid repeatedly looking up the column
descriptor during scanning. makeIndexKeyVals() is used to create a
[]parser.Datum from an index descriptor and populate it with datums of
the correct type.

See #2047.