Rework privilege storage.
Work towards #2005.

Credentials are now stored as a sorted list of {user, privilegeBitField}
The ALL logic makes grant/revoke a bit twisted, but is probably the most
expected behavior.

GRANT, REVOKE, and SHOW GRANTS are finalized.

The final step will be to fix all the CheckPrivilege calls to use the proper
privilege and finally remove READ,WRITE.