Add a row sentinel.
Every row has a sentinel that is the primary index key. The sentinel is
created when a row is inserted and deleted with the row.

Don't output keys for the primary key column values. The primary key
column values are already encoded in the primary key and scanNode was
never using the value.