Separate heartbeat failure from cleanup in TxnCoordSender.
If a heartbeat fails, stop heartbeating but do not perform
other cleanup. The client doesn't know the heartbeat
has failed and so will still try to complete the transaction,
and it needs to get a retryable error instead of the util.Errorf
that would result if the heartbeat removed the transaction.

Closes #2028.