Add EXPLAIN (plan).
The output is pretty redumentary, containing Level, Type and Description
field. The Level field indicates the level in the query plan tree (with
0 indicating the root). The Type indicates the type of the plan
node (currently scan, revscan, sort, nosort, or values). The Description
is some node specificy information, such as the table/index being
scanned or the keys being sorted on.

Fixed a buglet in valuesNode: the values lists must all be the same
length.

Fixes #2172.