Add the actual content to a gossip callback.
The typical pattern for a gossip callback method will be to perform
some action after receiving the update content. Currently, it will
only receive the updated key and not the content, requiring the
method to perform another locking call to the infostore. This removes
the need for the second call.