Increase tolerance of TestRetryExceedsMaxBackoff
This test flaked in a circle CI build, increasing the duration tolerance.

It was supposed to be verifying that the execution time was two orders of
magnitude faster than the worst-case scenario where MaxBackoff was not being
respected.  It was actually constraining to three orders of magnitude less (10e2
is 1000).