Implement allocator.RemoveTarget
allocator.RemoveTarget will be used to determine which replica in a given
replica set is the best candidate to be removed. The criteria is currently very
simple: the replica with the least available capacity is chosen for removal.

This will be used by the replicate queue to remove replicas from ranges that are
over-replicated.