Tiny fix: Make a header key same as a transaction key in Admin{Split,Merge}
For a txn created from AdminSplit, its txn key is created from the
start key of the second half of split. It does not always match with
Args.Key (e.g., the second admin split request in splitQueue.process).

For a txn created from AdminMerge, a txn key is created from the start
key of a subsuming range.

TxnCoordSender.sendOne() will anyway override a header key with
a transaction key, but it would be nice to have the matching
value for better readability.