Gossip SystemDB span.
Work towards #2179

Gossip the entire SystemDB span when a replica containing it is
added, but mainly when the new EndTransactionRequest trigger is set.
This is to avoid having to scan all keys in a transaction.

The sql side does need to be careful and set it everywhere it mutates
the system DB and tables. This is currently done by the addition of
a function filter in sql/main_test which fails EndTransaction requests
that have intents for SystemDB keys, but no trigger.

The setting is persisted in the session to be able to handle a sql
transaction that first mutates the systemDB, then performs a normal op.

Note: we could technically skip the trigger entirely, and use the field
on the transaction proto. This would deviate from the existing triggers,
but would keep things a bit cleaner and shorter.