Fix parsing of math.MinInt64.
Parse integer tokens using strconv.ParseUint, but only allow values in
the range [0, 9223372036854775808]. The resulting value is then cast to
an int64 resulting in values in the range [0, 9223372036854775807] +
-9223372036854775808. During expression normalization, if we see unary
minus of the value -9223372036854775808 we know that we really parsed
9223372036854775808 and are now negating it. During expression
evaluation, if we see a negative IntVal we know that we're attempting to
use (positive) 9223372036854775808 which is outside of the valid int64
range.

Fixes #2302.