Fix TODOs about Subquery and ExistsExpr in WalkExpr.
Fixed a bug in simplifyExpr where the expression "false" would simplify to
"true".

Added handling of "constant" filter expressions to index selection. If
the filter expression is the constant "false", we avoid scanning the
table.

Update the SQL used in the bank example and benchmark to only allow
transfers when a non-negative balance would result.