Move all replica adjustment logic into Allocator
This commit refactors all of the significant computations in the replicate queue
into the Allocator class. The replicate queue is now responsible only for
obtaining the necessary descriptors (zone config and range descriptor), passing
them to the allocator, and performing any actions recommended by the allocator.

This refactoring has an eye towards the removal of dead replicas; however, dead
replicas are not currently considered, as that functionality is not part of
storePool as of this commit.

This creates a very simple interface to all of the important logic, which can be
easily consumed by our planned allocation simulator.