Add Rebalancing to Replicate Queue
ReplicateQueue will now attempt to rebalance replicas to other stores in some
cases.

+ If the ReplicateQueue's `shouldQueue()` detects that no action is *necessary* on a particular
replica, it calls `allocator.shouldRebalance()` to determine if a rebalancing
opportunity might be wise.
+ If the ReplicateQueue's `process()` method encounters a replica for which no
action is *necessary*, it calls `allocator.rebalanceTarget()`. If that method
returns a store, a replica is added on that store.

Rebalancing must be enabled via the `RebalancingOptions` on StoreContext;
currently, this is set to false on the default server, and is only enabled for
the rebalancing test.

Furthermore, typical Rebalance operations include some randomness in decisions
in the anticipation of distributing load better when multiple stores are trying
to rebalance. This randomness can be disabled via the `Deterministic` option,
also on the RebalancingOptions structure.

This commit modifies multiTestContext's rpcSend method to better address the
situation where replicas are being changed.