Trim the grammar.
Removed support for typed tables (CREATE TABLE ... OF <type>) which take
their structure from a composite type. We've already removed support for
custom types, so this was just detritus.

Removed support for the INHERITS, WITH and ON COMMIT clauses to CREATE
TABLE. It is possible these will be reconstituted at some point, but
their implementation is not in any near term roadmap.

Removed support for the generalized forms of "ALTER TABLE COLUMN x
{SET,RESET}". Leave support for SET NULL and SET NOT NULL.

Removed support for a few other "ALTER TABLE" commands that we're
unlikely to support ("SET {LOGGED,UNLOGGED}", etc).

Removed support for "CREATE TABLE ... AS <query>". This can be
accomplished by creating the table and populating it with "INSERT INTO
<table> <query>".

Removed support for the RETURNING clause (a postgres extension) for
delete, insert and update statements.

Removes support for "ALTER DATABASE name" (which seemed to be useless)
and "ALTER DATABASE name SET var = value". We don't have any use for the
latter syntax at this point and can reconstitute in the future if
needed.

See #1949.