Remove support for the locking clause from SELECT.
The locking clause (FOR UPDATE, FOR SHARE, etc) allows the specification
of how rows are locked during select processing as they are obtained
from the table. These clauses don't make sense for Cockroach given the
lack of row-level locks.

Should we keep these clauses around and either ignore them or return an
error when they are specified?