Change PushTxnRequest to accept an optional PusherTxn.
This removes the dependency on RequestHeader.Txn, which was a bad practice
and potentially problematic as it's not semantically correct. The request
header txn value is meant always to indicate direct client operations using
a txn, and this isn't the case for pushes. Amongst other things, the request
to push will go through the coordinator local to the node encountering the
intent, which is not necessarily the coordinator which the client is bound
to for the txn.

Fixes #2300.