Stricter enforcement of splitable ranges in queues.
Fixes #2499

Add a new queue implementation method: `acceptsUnsplitRanges`
If false, ranges are checked for needed splits due to zone configs
(zone boundaries only, not max size).
The check is performed both before `shouldQueue` and `process`.

Since we're loading the system config, pass it along to the
`shouldQueue` and `process` methods. Some queues use it, and we want
to make sure we work using the same information.

Queues that accept ranges that need splitting:
* range_gc_queue
* split_queue (obviously)
* verify_queue

Queues that don't accept ranges that need splitting.
* gc_queue (looks up GC policy)
* replicate_queue (looks up attributes)

The only difference with previous behavior is in the replicate_queue's
`process` method which did not bail out on ranges that needed splitting,
although its `shouldQueue` method did check.