Add unique_id() builtin.
Unique IDs are composed of the current time in milliseconds, a 31-bit
random number and a 32-bit node-id. If two unique IDs are generated
within the same millisecond on a server, the random number is
incremented instead of being generated fresh. This ensures that unique
ID generation is monotonic on a single server and k-sorted across
servers.

Fixes #2513.