Add type info to proto.Value.
Changed Value.Tag from a string to a ValueType enum. Added INT, FLOAT,
BYTES and TIME value types and setters and getters for values of those
types.

Removed some of the unused flexibility from the "client"
package. Removes support for passing fmt.Stringer as the key to various
methods. Removed support for encoding.BinaryMarsaler.

Simplified the client.KeyValue structure. It now contains a Key and
Value where the Value is a *proto.Value. The Timestamp field has been
replaced with a Timestamp() method.

Fixes #2478.