clone args.Txn when copying into reply.Txn
fixes #2537 (or so I think). I think the race stems from an EndTransaction
which doesn't find a record on disk, using a shallow copy of args.Txn as
reply.Txn, which then gets set into a TransactionRetryError which can be
accessed out-of-band while args.Txn is still being used for retries.