remove errWithIndex
remove errWithIndex, which was used in `./storage` to wrap a batch
execution error along with the position at which it occurred.
That construct was prone to bugs and the later unwrapping into
a proto.Error meant that you could end up with a proto.Error
which carried more information than its associated ErrorDetail
(for example, `errWithIndex{err: ConditionalPutError, index:1}`
would translate into `proto.Error{Index: 1, Detail: ConditionalPutError}`
but since `ConditionalPutError` is a valid `ErrorDetail`, `GoError()`
would then return exactly that, losing the index.

Now an error whose relative position could be relevant at some future
point needs to implement `proto.IndexedError`, and then everything will
work as expected, regardless of whether the original error is an
`ErrorDetail` or not.