Work towards using MIN/MAX for index selection.
Compute desired ordering when aggregation functions are being used. If
there is a single MIN(k) or MAX(k) aggregate function, the desired
ordering is set to that column.

During ordering analysis for index selection, compute the prefix of the
index for which an exact match has been requested. This prefix is
ignored when determining how much of the index ordering matches the
requested ordering.

See #2636.