Simplify IS NULL and IS NOT NULL expressions.
Use "a != <val>" and "a IS NOT NULL" expressions to restrict the start
of scan ranges to not include NULL values. Use "a IS NULL" expressions
to restrict the end range to stop at the end of the NULL values.

The code is thorny. I recommend concentrating on the test cases.

See #2636.