disable txn auto gc
this could lead to spurious test failures:

* txn writes some intents, but not in a single
  round-trip
* concurrent reader runs into open intent, tries to push
* before it gets there, txn finishes with only local intents;
  txn record gets deleted
* push begins, does not find txn entry and fails
* potentially repeat forever or (if a push succeeds) until
  recreated txn entry times out.

I ran into this since it created an endless loop in
conjunction with buggy test code, but it's better to
disable this for now until we actually have the GC
story in place. My intuition is that we'll want to
limit this to single-roundtrip txns, whose intents
are never visible to anyone except the requests in
the batch (and even then we must be careful that
they are not collected as "skipped intents" in
some cases).