Enhance simplifyExpr to track equivalance of the returned expression.
Apply index constraints to the WHERE clause in order to strip out
portions of the filter which are guaranteed to be true.

Add a "col IS NOT NULL" constraint to the WHERE clause if the groupNode
has a desired ordering on "col". This currently occurs if there is a
single "{MIN,MAX}(col)" aggregation function.

Limit the number of keys scanned to 1 if the desired group-by ordering
matches the index ordering and the WHERE clause expression has been
reduced to "true".

Fixes #2636.