Change group.waitForCallback to a counter
I don't know exactly what happened inside Raft, but when I ran
TestRemoveLeader in multiraft_test.go, there was a writeResponse
that contains two committed ConfChanges (one for adding new node 4 and the
other for removing node 0). The writeResponse itself was sent the
newly added node 4.

To handle such a case correctly, group.waitForCallback needs to be a
count, not a bool.